# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Liquidity Hunter — SMC Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SymbolInfo {
  symbol: string;
  label: string;
  market: string;
  /** @nullable */
  correlatedSymbol?: string | null;
}

export interface SymbolList {
  crypto: SymbolInfo[];
  forex: SymbolInfo[];
}

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface StructurePoint {
  index: number;
  price: number;
  type: string;
  confirmed: boolean;
  time: number;
}

export interface StructureBreak {
  index: number;
  price: number;
  type: string;
  direction: string;
  time: number;
}

export interface StructureResult {
  trend: string;
  bias: string;
  confidence: number;
  pivots: StructurePoint[];
  breaks: StructureBreak[];
}

export interface LiquidityPool {
  price: number;
  type: string;
  score: number;
  touches: number;
  wasSwept: boolean;
  /** @nullable */
  sweptAt?: number | null;
  time: number;
  index: number;
  /** @nullable */
  session?: string | null;
}

export interface LiquidityResult {
  pools: LiquidityPool[];
  nearestBSL?: LiquidityPool | null;
  nearestSSL?: LiquidityPool | null;
}

export interface OrderBlock {
  type: string;
  proximal: number;
  distal: number;
  time: number;
  index: number;
  valid: boolean;
  isMitigated?: boolean;
  isBreaker?: boolean;
  strength: number;
  hasFvg: boolean;
}

export interface FairValueGap {
  type: string;
  top: number;
  bottom: number;
  time: number;
  index: number;
  fillFraction: number;
  isInversion?: boolean;
}

export interface DealingRange {
  high: number;
  low: number;
  timeframe: string;
}

export interface PdZone {
  label: string;
  top: number;
  bottom: number;
  timeframe: string;
  type: string;
}

export interface PdArrayResult {
  currentBias: string;
  zones: PdZone[];
  dealingRange: DealingRange;
  equilibrium: number;
}

export interface DailyBiasResult {
  bias: string;
  strength: number;
  consecutiveDays: number;
  /** @nullable */
  referencedSwing?: string | null;
}

export interface SmtDivergence {
  detected: boolean;
  /** @nullable */
  type?: string | null;
  confidence: number;
  /** @nullable */
  time?: number | null;
  /** @nullable */
  primarySymbol?: string | null;
  /** @nullable */
  correlatedSymbol?: string | null;
}

export interface DrawTarget {
  price: number;
  type: string;
  score: number;
  direction: string;
  label: string;
}

export interface SmcReport {
  symbol: string;
  market: string;
  timeframe: string;
  currentPrice: number;
  generatedAt: number;
  candles: Candle[];
  structure: StructureResult;
  liquidity: LiquidityResult;
  orderBlocks: OrderBlock[];
  fvg: FairValueGap[];
  pdArray: PdArrayResult;
  dailyBias: DailyBiasResult;
  smt?: SmtDivergence;
  draw: DrawTarget[];
}

export type AnalyzeCryptoParams = {
/**
 * Binance symbol e.g. BTCUSDT
 */
symbol: string;
timeframe?: string;
/**
 * Symbol for SMT divergence e.g. ETHUSDT
 */
correlatedSymbol?: string;
};

export type AnalyzeForexParams = {
/**
 * Yahoo Finance symbol e.g. EURUSD=X
 */
symbol: string;
timeframe?: string;
/**
 * Yahoo symbol for SMT divergence e.g. GBPUSD=X
 */
correlatedSymbol?: string;
};