# lib/api-zod/src/generated/types/orderBlock.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Liquidity Hunter — SMC Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderBlock {
  type: string;
  proximal: number;
  distal: number;
  time: number;
  index: number;
  valid: boolean;
  isMitigated?: boolean;
  isBreaker?: boolean;
  strength: number;
  hasFvg: boolean;
}