# lib/api-zod/src/generated/types/smtDivergence.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Liquidity Hunter — SMC Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface SmtDivergence {
  detected: boolean;
  /** @nullable */
  type?: string | null;
  confidence: number;
  /** @nullable */
  time?: number | null;
  /** @nullable */
  primarySymbol?: string | null;
  /** @nullable */
  correlatedSymbol?: string | null;
}