# lib/api-zod/src/generated/types/structurePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Liquidity Hunter — SMC Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface StructurePoint {
  index: number;
  price: number;
  type: string;
  confirmed: boolean;
  time: number;
}