# lib/api-zod/src/generated/types/symbolInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Liquidity Hunter — SMC Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface SymbolInfo {
  symbol: string;
  label: string;
  market: string;
  /** @nullable */
  correlatedSymbol?: string | null;
}